# src/docs/swagger.yaml
definitions:
...
  github_com_wahid-wex_e-commerce-backend_api_dto.CartItemRequest:
    properties:
      product_id:
        type: integer
    required:
    - product_id
    type: object
  github_com_wahid-wex_e-commerce-backend_api_dto.CartItemResponse:
    properties:
      imageUrl:
        type: string
      name:
        type: string
      price:
        type: number
      productId:
        type: integer
      quantity:
        type: integer
    type: object
  github_com_wahid-wex_e-commerce-backend_api_dto.CartResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.CartItemResponse'
        type: array
      totalPrice:
        type: number
    type: object
  github_com_wahid-wex_e-commerce-backend_api_dto.CategoryResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      imageUrl:
        type: string
      name:
        type: string
    type: object
...
  github_com_wahid-wex_e-commerce-backend_api_dto.CreateUpdateCategoryRequest:
    properties:
      description:
        maxLength: 300
        minLength: 10
        type: string
      imageUrl:
        maxLength: 100
        minLength: 3
        type: string
      name:
        maxLength: 20
        minLength: 3
        type: string
    required:
    - description
    - imageUrl
    - name
    type: object
...
  github_com_wahid-wex_e-commerce-backend_api_dto.Filter:
    properties:
      filterType:
        description: text number
        type: string
      from:
        type: string
      to:
        type: string
      type:
        type: string
    type: object
...
  github_com_wahid-wex_e-commerce-backend_api_dto.PagedList-github_com_wahid-wex_e-commerce-backend_api_dto_CategoryResponse:
    properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      items:
        items:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.CategoryResponse'
        type: array
      pageNumber:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
...
  github_com_wahid-wex_e-commerce-backend_api_dto.PaginationInputWithFilter:
    properties:
      filter:
        additionalProperties:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.Filter'
        type: object
      pageNumber:
        type: integer
      pageSize:
        type: integer
      sort:
        items:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.Sort'
        type: array
    type: object
...
  github_com_wahid-wex_e-commerce-backend_api_dto.Sort:
    properties:
      colId:
        type: string
      sort:
        type: string
    type: object
  github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse:
    properties:
      error: {}
      result: {}
      resultCode:
        $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.ResultCode'
      success:
        type: boolean
      validationErrors:
        items:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_validation.ValidationError'
        type: array
    type: object
  github_com_wahid-wex_e-commerce-backend_api_helper.ResultCode:
    enum:
    - 0
    - 40001
    - 40101
    - 40301
    - 42901
    - 50001
    - 50002
    type: integer
    x-enum-varnames:
    - Success
    - ValidationError
    - AuthError
    - ForbiddenError
    - LimiterError
    - CustomRecovery
    - InternalError
  github_com_wahid-wex_e-commerce-backend_api_validation.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
info:
  contact: {}
paths:
  /v1/cart/:
    delete:
      consumes:
      - application/json
      description: Remove From Cart
      parameters:
      - description: Remove Product to Cart
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.CartItemRequest'
      responses:
        "201":
          description: Product response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.CartResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Remove From Cart
      tags:
      - Cart
    get:
      consumes:
      - application/json
      description: Get Cart Detail
      responses:
        "200":
          description: Cart Response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.CartResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get Cart
      tags:
      - Cart
    post:
      consumes:
      - application/json
      description: Add Product to Cart
      parameters:
      - description: Add Product to Cart
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.CartItemRequest'
      responses:
        "201":
          description: Product response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.CartResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Add Product to Cart
      tags:
      - Cart
  /v1/categories/:
    post:
      consumes:
      - application/json
      description: Create a Category
      parameters:
      - description: Create a Category
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.CreateUpdateCategoryRequest'
      responses:
        "201":
          description: Category response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.CategoryResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
      summary: Create a Category
      tags:
      - Categories
  /v1/categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Category
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
      summary: Delete a Category
      tags:
      - Categories
    get:
      consumes:
      - application/json
      description: Get a Category
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Category response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.CategoryResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
      summary: Get a Category
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Update a Category
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update a Category
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.CreateUpdateCategoryRequest'
      responses:
        "200":
          description: Category response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.CategoryResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
      summary: Update a Category
      tags:
      - Categories
  /v1/categories/get-by-filter:
    post:
      consumes:
      - application/json
      description: Get a Category
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.PaginationInputWithFilter'
      responses:
        "200":
          description: Category response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.PagedList-github_com_wahid-wex_e-commerce-backend_api_dto_CategoryResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
      summary: Get a Category
      tags:
      - Categories
...
securityDefinitions:
  AuthBearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"